# labels: Use slices.Sort instead of sort.Strings